refactor: SseEmitter 관리 코드 리팩토링 (Provider -> Repository, Util로 분리)
