+++ src/main/java/com/bitharmony/comma/album/streaming/util/EncodeStatus.java
@@ -1,6 +1,7 @@
 package com.bitharmony.comma.album.streaming.util;
 
 public enum EncodeStatus {
+    CONNECT,
     COMPLETE,
     WAITING,
     RUNNING,
