+++ src/main/java/com/bitharmony/comma/album/streaming/util/SseEmitterUtil.java
@@ -17,9 +17,13 @@ public class SseEmitterUtil {
     private final SseEmitterRepository sseEmitterRepository;
     private final static Long DEFAULT_TIMEOUT = 60000L * 10;
     private final static Long RECONNECT_TIME = 30000L;
+    private final static String EVENT_NAME = "Encoding Status";
 
-    public SseEmitter createSseEmitter() {
-        return new SseEmitter(DEFAULT_TIMEOUT);
+    public SseEmitter generateSseEmitter(String filePath) {
+        SseEmitter sseEmitter = new SseEmitter(DEFAULT_TIMEOUT);
+        sseEmitterRepository.save(filePath, sseEmitter);
+        sendEvent(filePath, EVENT_NAME, EncodeStatus.CONNECT);
+        return sseEmitter;
     }
 
     public void sendEvent(String key, Object... event) {
@@ -30,29 +34,31 @@ public class SseEmitterUtil {
             sseEmitter.send(SseEmitter.event().reconnectTime(RECONNECT_TIME)
                     .data(event, MediaType.APPLICATION_JSON)
                     .id(UUID.randomUUID().toString())
-                    .name("Notification")
+                    .name(EVENT_NAME)
             );
 
         } catch (IOException e) {
             throw new SseEmitterSendingException();
         }
-    }
 
-    public void sendEventWithSseEmitter(SseEmitter sseEmitter , Object... event) {
-        try {
-            sseEmitter.send(SseEmitter.event().reconnectTime(RECONNECT_TIME)
-                    .data(event, MediaType.APPLICATION_JSON)
-                    .id(UUID.randomUUID().toString())
-                    .name("Notification")
-            );
-        } catch (IOException e) {
-            throw new SseEmitterSendingException();
+        if (isComplete(event)) {
+            complete(sseEmitter, key);
         }
     }
 
     public void complete(SseEmitter sseEmitter, String key) {
         sseEmitter.complete();
         sseEmitterRepository.deleteByKey(key);
+    }
+
+    private boolean isComplete(Object... events) {
+        for (Object event : events) {
+            if (event == EncodeStatus.COMPLETE) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
 }
